+++ test/integration/geo_search_integration_test.rb
@@ -21,10 +21,32 @@
 require 'helper'
 
 describe "RiakJson provides geospatial Solr search" do
+  # The '_rj-rb-client-us-capitals' docs were inserted in test/seeds/seeds.rb
+  
   it "ensure seeding of us capitals worked" do
     client = rj_test_client
     collection = client.collection('_rj-rb-client-us-capitals')
     results = collection.all()
     results.wont_be_empty
+  end
+  
+  it "can perform arbitrary geo Solr queries" do
+    client = rj_test_client
+    collection = client.collection('_rj-rb-client-us-capitals')
+    spatial_field = 'capital_coords_rpt'
+    location = '41.82355,-71.422132'  # Providence, RI
+    distance = '300' # kilometers
+    filter = true  # exclude results that are outside of the distance?
+    query_params = {
+      :fl => '*,score',  # field list
+      :sort => 'score asc',
+      :q => "{!geofilt score=distance filter=#{filter} sfield=#{spatial_field} pt=#{location} d=#{distance}}",
+      :wt => 'json'
+    }
+    # Get a raw Solr result in JSON format back
+    result = collection.solr_query_raw(query_params)
+    result = JSON.parse(result)
+#    puts JSON.pretty_generate(result)
+    result['response']['numFound'].must_equal 6  # There should be 6 state capitals within 300km of Providence RI
   end
 end